+++ pytket/docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 Minor new features:
 * New properties: :py:meth:``circuit.Op.dagger`` and :py:meth:``circuit.Op.transpose``.
+* New methods: :py:meth:``routing.Placement.to_dict`` and :py:meth:``routing.Placement.from_dict``.
 
 0.16.0 (October 2021)
 ---------------------